requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-experimental>=0.3.0
openai>=1.40.0

# GitHub access
PyGithub>=2.3.0

# Environment variables (.env file)
python-dotenv>=1.0.1

# PDF generation
pylatex>=1.4.2

# Typing and runtime helpers
typing-extensions>=4.9.0
